[Fix] #9 MedicineSchedule 유니크 제약조건 추가 및 Medicine 삭제 시 고아 여부 확인

+++ src/main/java/com/ssu/ongi/domain/medicine/service/MedicineScheduleCommandService.java
@@ -85,6 +85,9 @@ public class MedicineScheduleCommandService {
         medicationRecordRepository.deleteAllByMedicineScheduleId(scheduleId);
         Medicine medicine = schedule.getMedicine();
         medicineScheduleRepository.delete(schedule);
-        medicineRepository.delete(medicine);
+
+        if (!medicineScheduleRepository.existsByMedicineId(medicine.getId())) {
+            medicineRepository.delete(medicine);
+        }
     }
 }
